requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
filelock==3.16.1
fsspec==2024.10.0
huggingface-hub==0.26.3
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
packaging==24.2
pillow==11.0.0
protobuf==5.29.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
sentencepiece==0.2.0
setuptools==75.6.0
sympy==1.13.1
tiktoken==0.8.0
tokenizers==0.20.3
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
urllib3==2.2.3